Use a direct call to get the container rather than paging through the list.
There's a bit in the API documentation that makes it seem like we coudl get an empty page _with_ a 'next' marker. This could mean we'd have to call _back_ intot he paging API in a loop.

Rather than doing that I'll just grab the container info directly and check for 404.